+++ Sources/StreamChatSwiftUI/ChatChannel/MessageList/MessageListConfig.swift
@@ -27,7 +27,10 @@ public struct MessageListConfig {
         showNewMessagesSeparator: Bool = true,
         handleTabBarVisibility: Bool = true,
         messageListAlignment: MessageListAlignment = .standard,
-        uniqueReactionsEnabled: Bool = false
+        uniqueReactionsEnabled: Bool = false,
+        localLinkDetectionEnabled: Bool = true,
+        isMessageEditedLabelEnabled: Bool = true,
+        markdownSupportEnabled: Bool = true
     ) {
         self.messageListType = messageListType
         self.typingIndicatorPlacement = typingIndicatorPlacement
@@ -48,6 +51,9 @@ public struct MessageListConfig {
         self.handleTabBarVisibility = handleTabBarVisibility
         self.messageListAlignment = messageListAlignment
         self.uniqueReactionsEnabled = uniqueReactionsEnabled
+        self.localLinkDetectionEnabled = localLinkDetectionEnabled
+        self.isMessageEditedLabelEnabled = isMessageEditedLabelEnabled
+        self.markdownSupportEnabled = markdownSupportEnabled
     }
 
     public let messageListType: MessageListType
@@ -69,6 +75,9 @@ public struct MessageListConfig {
     public let handleTabBarVisibility: Bool
     public let messageListAlignment: MessageListAlignment
     public let uniqueReactionsEnabled: Bool
+    public let localLinkDetectionEnabled: Bool
+    public let isMessageEditedLabelEnabled: Bool
+    public let markdownSupportEnabled: Bool
 }
 
 /// Contains information about the message paddings.